Restructure of core module and addition of aspect class to check api response times.
