+++ hotel-core/hotel-core-app/src/main/java/gr/balasis/hotel/core/app/base/BaseComponent.java
@@ -1,0 +1,20 @@
+package gr.balasis.hotel.core.app.base;
+
+import jakarta.annotation.PostConstruct;
+import jakarta.annotation.PreDestroy;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public abstract class BaseComponent {
+    protected Logger logger = LoggerFactory.getLogger(getClass());
+
+    @PostConstruct
+    public void init() {
+        logger.trace("Loaded {}.", getClass());
+    }
+
+    @PreDestroy
+    public void destroy() {
+        logger.trace("{} is about to be destroyed.", getClass().getName());
+    }
+}
