+++ hotel-core/hotel-core-app/src/main/java/gr/balasis/hotel/core/app/config/CoreAppConfig.java
@@ -1,0 +1,9 @@
+package gr.balasis.hotel.core.app.config;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan(basePackages = "gr.balasis.hotel.core.monitoring")
+public class CoreAppConfig {
+}
