+++ hotel-core/hotel-core-app/src/main/java/gr/balasis/hotel/core/app/config/FeedbackModuleConfiguration.java
@@ -1,0 +1,16 @@
+package gr.balasis.hotel.core.app.config;
+
+
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@Configuration
+@Profile("dev-feedback")
+@ComponentScan(basePackages = "gr.balasis.hotel.modules.feedback")
+@EntityScan(basePackages = "gr.balasis.hotel.modules.feedback.entity")
+@EnableJpaRepositories(basePackages = "gr.balasis.hotel.modules.feedback.repository")
+public class FeedbackModuleConfiguration {
+}
