+++ hotel-core/hotel-core-app/src/main/java/gr/balasis/hotel/core/app/config/MessageSourceConfig.java
@@ -1,0 +1,18 @@
+package gr.balasis.hotel.core.app.config;
+
+import org.springframework.context.MessageSource;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.support.ReloadableResourceBundleMessageSource;
+
+@Configuration
+public class MessageSourceConfig {
+
+    @Bean
+    public MessageSource messageSource() {
+        ReloadableResourceBundleMessageSource messageSource = new ReloadableResourceBundleMessageSource();
+        messageSource.setBasename("classpath:feedback-messages");
+        messageSource.setDefaultEncoding("UTF-8");
+        return messageSource;
+    }
+}
