+++ hotel-core/hotel-core-app/src/main/java/gr/balasis/hotel/core/app/service/BaseService.java
@@ -1,0 +1,11 @@
+package gr.balasis.hotel.core.app.service;
+
+import java.util.List;
+
+public interface BaseService <T,K>{
+    T create(final T item);
+    void update(T item);
+    void delete(T item);
+    T findById(K id);
+    List<T> findAll();
+}
