+++ hotel-core/hotel-core-app/src/main/java/gr/balasis/hotel/core/app/service/RoomService.java
@@ -1,0 +1,7 @@
+package gr.balasis.hotel.core.app.service;
+
+import gr.balasis.hotel.context.base.domain.domains.Room;
+
+public interface RoomService extends BaseService<Room,Long>{
+    boolean existsById(Long guestId);
+}
