+++ hotel-core/hotel-core-app/src/main/resources/application-dev-feedback.yml
@@ -1,0 +1,33 @@
+spring:
+  datasource:
+    url: jdbc:h2:mem:devDB
+    driver-class-name: org.h2.Driver
+    username: feedbackUser
+    password: feedbackPass
+    hikari:
+      maximum-pool-size: 4
+      minimum-idle: 2
+      idle-timeout: 500000
+      max-lifetime: 1200000
+      connection-timeout: 30000
+
+  jpa:
+    hibernate:
+      ddl-auto: update
+    properties:
+      hibernate:
+        dialect: org.hibernate.dialect.H2Dialect
+
+  h2:
+    console:
+      enabled: true
+      path: /h2-console
+      settings:
+        web-allow-others: true
+
+logging:
+  level:
+    root: info
+    gr.balasis.hotel.core.app: trace
+    gr.balasis.hotel.modules.feedback: trace
+    gr.balasis.hotel.core.monitoring: trace
