+++ hotel-core/hotel-core-app/src/main/resources/application.yml
