+++ hotel-core/hotel-core-app/src/main/resources/feedback-messages.properties
