--- hotel-core/src/main/java/gr/balasis/hotel/core/config/MessageSourceConfig.java
@@ -1,18 +1,0 @@
-package gr.balasis.hotel.core.config;
-
-import org.springframework.context.MessageSource;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.support.ReloadableResourceBundleMessageSource;
-
-@Configuration
-public class MessageSourceConfig {
-
-    @Bean
-    public MessageSource messageSource() {
-        ReloadableResourceBundleMessageSource messageSource = new ReloadableResourceBundleMessageSource();
-        messageSource.setBasename("classpath:feedback-messages");
-        messageSource.setDefaultEncoding("UTF-8");
-        return messageSource;
-    }
-}
