--- hotel-core/src/main/java/gr/balasis/hotel/core/service/RoomService.java
@@ -1,7 +1,0 @@
-package gr.balasis.hotel.core.service;
-
-import gr.balasis.hotel.context.base.domain.domains.Room;
-
-public interface RoomService extends BaseService<Room,Long>{
-    boolean existsById(Long guestId);
-}
