+++ pom.xml
@@ -18,9 +18,10 @@
 
     <modules>
         <module>hotel-context</module>
-        <module>hotel-core</module>
         <module>hotel-modules</module>
         <module>hotel-data</module>
+        <module>hotel-core</module>
+
     </modules>
 
     <properties>
